feat: Adds initial point store size parameter
Adds the `initial_point_store_size` parameter to the Random Cut Forest model.

This parameter allows users to configure the initial size of the
point store, potentially improving performance and memory usage
depending on the data characteristics. If not provided, a default
value of twice the sample size is used.